Cleaned-up Windows code and moved common code to a header file
