+++ test/test_write.c
@@ -8,66 +8,7 @@ This file is part of H5Z-ZFP. Please also read the BSD license
 https://raw.githubusercontent.com/LLNL/H5Z-ZFP/master/LICENSE 
 */
 
-#ifndef _GNU_SOURCE
-#define _GNU_SOURCE
-#endif
-#include <assert.h>
-#include <errno.h>
-#include <fcntl.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-#if defined(_WIN32) || defined(_WIN64)
-#define _USE_MATH_DEFINES
-#include <math.h>
-#define j0 _j0
-#include <io.h>
-#define strncasecmp _strnicmp
-#define strcasecmp _stricmp
-#define read _read
-#define open _open
-#define close _close
-#define srandom(X) srand(X)
-#define random rand
-
-// strcasestr is not available on Windows
-#include <ctype.h>
-char * strcasestr(s, find)
-     const char *s, *find;
-{
-  char c, sc;
-  size_t len;
-
-  if ((c = *find++) != 0) {
-    c = tolower((unsigned char)c);
-    len = strlen(find);
-    do {
-      do {
-        if ((sc = *s++) == 0)
-          return (NULL);
-      } while ((char)tolower((unsigned char)sc) != c);
-    } while (strncasecmp(s, find, len) != 0);
-    s--;
-  }
-  return ((char *)s);
-}
-
-// strndup() is not available on Windows
-char *strndup( const char *s1, size_t n)
-{
-    char *copy= (char*)malloc( n+1 );
-    memcpy( copy, s1, n );
-    copy[n] = 0;
-    return copy;
-};
-
-#else
-#include <math.h>
-#include <unistd.h>
-#endif
-
-#include "hdf5.h"
+#include "test_common.h"
 
 #ifdef H5Z_ZFP_USE_PLUGIN
 #include "H5Zzfp_plugin.h"
@@ -85,96 +26,6 @@ char *strndup( const char *s1, size_t n)
   #error GOT HERE
 #endif
 #endif
-
-#define NAME_LEN 256
-
-/* convenience macro to handle command-line args and help */
-#define HANDLE_SEP(SEPSTR)                                      \
-{                                                               \
-    char tmpstr[64];                                            \
-    int len = snprintf(tmpstr, sizeof(tmpstr), "\n%s...", #SEPSTR);\
-    printf("    %*s\n",60-len,tmpstr);                          \
-}
-
-#define HANDLE_ARG(A,PARSEA,PRINTA,HELPSTR)                     \
-{                                                               \
-    int i;                                                      \
-    char tmpstr[64];                                            \
-    int len;                                                    \
-    int len2 = strlen(#A)+1;                                    \
-    for (i = 0; i < argc; i++)                                  \
-    {                                                           \
-        if (!strncmp(argv[i], #A"=", len2))                     \
-        {                                                       \
-            A = PARSEA;                                         \
-            break;                                              \
-        }                                                       \
-        else if (!strncmp(#A, "help", 4) &&                     \
-                  strcasestr(argv[i], "help"))                  \
-        {                                                       \
-            return 0;                                           \
-        }                                                       \
-    }                                                           \
-    len = snprintf(tmpstr, sizeof(tmpstr), "%s=" PRINTA, #A, A);\
-    printf("    %s%*s\n",tmpstr,60-len,#HELPSTR);               \
-}
-
-
-/* convenience macro to handle errors */
-#if defined(_WIN32) || defined(_WIN64)
-
-#define ERROR(FNAME)                                              \
-do {                                                              \
-    size_t errmsglen = 94;                                        \
-    char errmsg[errmsglen];                                       \
-    strerror_s(errmsg, errmsglen, errno);                         \
-    fprintf(stderr, #FNAME " failed at line %d, errno=%d (%s)\n", \
-            __LINE__, errno, errno?errmsg:"ok");                  \
-    return 1;                                                     \
-} while(0)
-
-#else
-
-#define ERROR(FNAME)                                              \
-do {                                                              \
-    int _errno = errno;                                           \
-    fprintf(stderr, #FNAME " failed at line %d, errno=%d (%s)\n", \
-        __LINE__, _errno, _errno?strerror(_errno):"ok");          \
-    return 1;                                                     \
-} while(0)
-
-#endif
-                                    \
-/* Generate a simple, 1D sinusioidal data array with some noise */
-#define TYPINT 1
-#define TYPDBL 2
-static int gen_data(size_t npoints, double noise, double amp, void **_buf, int typ)
-{
-    size_t i;
-    double *pdbl = 0;
-    int *pint = 0;
-
-    /* create data buffer to write */
-    if (typ == TYPINT)
-        pint = (int *) malloc(npoints * sizeof(int));
-    else
-        pdbl = (double *) malloc(npoints * sizeof(double));
-    srandom(0xDeadBeef);
-    for (i = 0; i < npoints; i++)
-    {
-        double x = 2 * M_PI * (double) i / (double) (npoints-1);
-        double n = noise * ((double) random() / ((double)(1<<31)-1) - 0.5);
-        if (typ == TYPINT)
-            pint[i] = (int) (amp * (1 + sin(x)) + n);
-        else
-            pdbl[i] = (double) (amp * (1 + sin(x)) + n);
-    }
-    if (typ == TYPINT)
-        *_buf = pint;
-    else
-        *_buf = pdbl;
-    return 0;
-}
 
 /* Populate the hyper-dimensional array with samples of a radially symmetric
    sinc() function but where certain sub-spaces are randomized through dimindx arrays */
